New FileCache class
Before, there was caching logic internally used within NetworkCache
which has now been promoted to its own, standalone class.

In this CL, we've refactored NetworkCache in a minor way to delegate the old
caching logic to the new class; however, in a later CL, the file cache
will be completely removed out.

Test: Basic usage tested. Some threading logic may need to be refactored
in such a way it can be tested later.

Bug: b/29578298

Change-Id: I2bba3a08f6aedc595e9e02663cdb4983409a21c4

Former-commit-id: 983605ce9aefcbc43477fe04ba8c4e24150dd862