Cherry-pick: dx changes
Original commit message:
Ignore LV debug info when the type doesn't match

As a consequence of the CL that relaxes the conditions to retrieve LV
debug info (455443f47d9e217b99edcc33202f889e2c637b7b), there may be a
SimException because of a "local variable type mismatch".
Some local variable slots may be used for different local variables
with different types, but only one has debug info.
So we need to check if the types match, and ignore the local variable
info if it doesn't.

Bug: 32432143
Test: manual
> Change-Id: I0a12ac6476f612c0d8e597c9690587826ac41841

Change-Id: Ib48c864ec5653eeff7b415c166ee125d187b1399

Former-commit-id: cd871ea1bdc91bd3774ec6a720f8d08c9b557e04