[AVD Mgr] Fix user-created round Wear devices
Interpret the diagonal size differently for round
devices than for square devices. For square, the
horizontal size = (diagonal / sqrt(2)), but for round
horizontal size = diagonal.

Remove a hack to Wear device names. That hack was
required previously, but breaks things now.

Ensure that a user-defined Wear device doesn't
forget if it should be round.

Test: Added DeviceManagerTest.testWriteUserDevice() and
      AvdScreenDataTest.testCalculateDpi()

BUG: b.android.com/200883

Change-Id: I878b84434c720f0f5698b87675682af750cbfd81

Former-commit-id: 59c9b491c9ca5f2b5720cf4255e3011c45d25fca