Fix dex incremental updates for packaging
If deleted .dex files updates were out of order, we
would not update DEX files properly, which would result
in NPE. Now, we are sorting list of deleted files, which
mitigates this issue.

Deleted files are treated differently because the packager
is trying to use all available names. So if there were
classes2.dex and classes3.dex, and we removed classes2.dex,
packager will move classes3.dex to classes2.dex.

Test: DexIncrementalRenameManagerTest
Change-Id: I45eea7e960865ba2096bc4ae3fe4e256cba52725

Former-commit-id: 614f945a2beba4c675de8a3dc2ce520a8d830870