devicelib: Add utility method to identify if adb server is running
This CL adds a utility method to check whether the adb server is already
running at a given port. It does this by attempting to connect to that
port and issuing an adb command ("get version"). If it can't connect, or
if the response doesn't look like it is from adb, then it returns false.

Test: unit tests included
Change-Id: Ie82ae7522c3185c19aaf863f780a121692e1eadf

Former-commit-id: 27a38be7a442b0c8f4c4c7f74517934f2284b50f