Reverting the builtin stats/perf from slicer
While the stats were invaluable during prototyping, their implementation
was very simplistic and intended only for the test version of dexter
(where we didn't have the separate slicer library)

We need to revisit the stats and come up with a more generic and safe
interface that's suitable for a library (slicer). In particular we need
thread (or even scope) local stats. Since this is a non-essential part
of the .dex instrumentation I'm leaving it out until we get the core
functionality in place.

Test: nothing to test

Change-Id: Iac3fac98a2a780778e0c74591af33d1c9cc49aba

Former-commit-id: 825059ce93a96ff449641f1a54fa142e9ce85e11