Fix unit tests with resources.
There was a missing task dependency with the asset merge
output. To fix this problem, this change moves the asset merge
output to the new FileCollection mechanism, and update all its
consumer to use the file collection.

Also updates the unit test code to depend on the content of the resource
and asset merger in order to rerun tests if only a res/asset change is
detected.

Test: fixing them.

Change-Id: Ia875185e1c5188c3978405f0e3362546fda0b3b4

Former-commit-id: 790c1452bfbf9ef1dfb54094c87b4d33f800cf5f