Provide safe file access across threads and processes.
When multiple threads or processes access the same file, they would
require some form of synchronization. The introduced SynchronizedFile
class provides a simple way for clients to add synchronization
capability to a file without having to work with low-level details
involving within-process or inter-process locking.

The implementation of this class is extracted from the existing
FileCache class, so that it can be used by any task that needs
file access synchronization.

An upcoming CL will use this class to make ValidateSigningTask
thread-safe. FileCache should also be refactored to use this new
class in another CL.

Test: SynchronizedFileTest
Change-Id: I1652720b090828e72b1a9e07361d0063866b5b27

Former-commit-id: 9e647730245eadb592c895b9c0e5ab4ee46ff048