Add lint check for instant apps
Work in progress: still needs a full issue explanation, more-info
links, and final error messages. Also, this is using a hardcoded
set of API limitations (such as whitelisted permissions etc) which
may need tweaking as the instant app SDK evolves.

Also, all issues are currently reported as a single lint issue
(with severity error); we may want to split it up into a warning-
severity issue and an error-severity issue.

Test: Unit tests included

Change-Id: I89c8bcdb05695e90e8fa8f35d64d0e88d7e35327

Former-commit-id: ec76b599827ce46c5a90e48b927d977a1c68845e