Enable legacy multi-dex with the built-in shrinker
Just make sure the jars are merged when necessary. Previously we didn't
want to do it this way, since shrinker's job is to improve performance
and jar-merging before dx is not good at that. But now we know
jar-merging is going away soon anyway, with the new, more incremental
dexing pipeline.

Test: MultiDexTest
Change-Id: I1d3b39e83f55791a7110334911822b675bec118a

Former-commit-id: 3cc04ffd0daec3a7f0e920f497306f046e2a6404