Fix creating cache in case creation fails.
If the creation fails with an exception, then the state
of the value (key) is not changed from "being-processed"
to non existant. This makes threads blocked on the creation
never return and threads trying to clear the cache later
fail with a weird error (values being processed)

Now also handles the case where a thread is query for a value
being processed, and the creation fails. Instead of returning
null, attempts to create it again. If more than one threads
is in this state, then they'll all race to attempt to be the
one creating it but the first one will win and the other ones
will be block on this new creation (it's using the existing
mechanism).

Test: new unit tests.
Change-Id: Ieeaddbc0bef5ac776d430d62a1c1e1ca6160a8a2

Former-commit-id: f68ddab3bbf41871b5d152bf3f107241524ee87a