Fix the model to include tested app via different modules.
Change the way we query for all the dependencies and
differentiate java vs android. We used to query for different
lists but this ommited non aar dependencies (test module
depending on a tested app module).

The new way queries for JAR which everything now uses,
and then separately for EXPLODED_AAR which only the android modules
use, and use that to disambiguate external dependencies.

With this, we also move the annotation processor to query for JAR
directly rather than for a new custom type.

Test: existing ones, plus tweaking the separate test ones.
Change-Id: Ia31203f623bc7dd4699e232d6740d23a730143a3

Former-commit-id: 40ba45e22c6e85638d75be1df928bac5839a8c2d