some cleaning, removed some old code.
InstantRun BuildContext was planned to be use as a general purpose build
context to give information about produced APKs and splits.
This decision was reverted and we now use an output.json file that's
specifically handle build outputs.

Removed some the craft added in the original CL and some unused code.

Test: removed tests for the dead code.
Change-Id: Ia064b4fb057e58751f06f2bccf9d382578a1400b

Former-commit-id: 7074176b50500aa3f3ce829d13117565bc33618f