Add file size limit on FileCache
Now, the file cache will limit itself to 50% of the available free space
on the device. This should help prevent perfd from accidentally
overloading the user's device with temporary files.

Writing tests to cover the file cache thread is not a trivial effort and
will have to wait for us to revisit integration testing of perfd. In the
meantime, I locally set the file cache limit very low and made sure that
older files were, indeed, getting removed first, and that the cache disk
limit was being honored.

Test: To come later
Bug: 34931701
Change-Id: I612519306c0962d0ab79569cd3414be6db630d49

Former-commit-id: f83c74fc1946de67f9ffa07a6945b474c40c6877