Lint: Warn about obsolete version folders, misc bug fixes
This CL fixes the following issues:
236018: Expand ObsoleteSdkInt lint check to include resources qualifiers
238476: Multidex Support Library Gradle Lint Error
237776: Lint CommitTransaction false positive if retval saved to local variable
230652: Incorrect lint error on instantiatable custom Views
234901: Build fails because of a missing class
236584: FullBackup Lint Warning Breaks with Merged Manifests
229664: Gradle 2.3.0-alpha2 lint check incorrectly flags projects with
        data binding and support libs

It also adds a new convenience class, QuickfixData, which makes it more
convenient to pass multiple pieces of data to the IDE for use by
quickfix implementations.

It also adds more mocking support for Gradle dependency graphs.

It also updates the lint testing infrastructure to require the SDK
to be present and available (not relevant for the builtin lint tests
but important for third party lint checks running outside the tools tree.)

It also adds the JetBrains Nullable/NotNull annotations to the do-not-
auto-import list such that they don't get accidentally used (in tools/base
we should be using the com.android.* ones.)

Test: Unit tests included
Change-Id: Ib862927793c9848f673308f30f303b0fc17f6293

Former-commit-id: 96b1c1c756e3a81bea96b6caa84fd55cd8cb5209