devicelib: Fix flaky OsProcessRunnerTest
The OsProcessRunner provides an API that allows users to wait until a
running process quits. When a process quits, the threads that are
reading the stdout and stderr streams may not yet be done.

This CL ensures that waiting for a process implies waiting for the
process output readers to also complete.

Test: existing test now passes even when run for 10k iterations, whereas
it would fail in < 100 iterations earlier
Change-Id: I1ce1dd2208cde542d0392635b929ec66658ecc53

Former-commit-id: 2a1b10e6e5b342a698bf55f5c1e72b806bf6fc3d