Add support for publishing/consuming type.
This adds the basic feature to have a consuming type, and
one or more publishing types. Nothing yet uses 2 publishing
types.

Note that this changes the behavior of other types (app, and
library). This makes a library unable to consume the content
of a library (which makes sense). Updated a test to reflect this
change in behavior.

Test: existing.

Change-Id: I9f4a63277ea231588041efecc3156282529dba38

Former-commit-id: 0c93cd34786b0bc8c768e3ed6cc9c2ba7c614bc5