Adding a callback to perfa for perfd state changes.
This change, adds a callback in perfa to inform subsystems of perfd state
changes. The current system that uses this is activities. The reason this
is required is in the scenaio studio crashes after a developer starts an
app with advanced profiling. On the second start of studio, the current
state of the activity world does not get sent across. This is because perfd
would have closed and cleared its cache. As well as the datastore would have
shutdown, and had its cache cleared.

Bug: 35919346
Test: None, we need to setup a test framework for this.

Change-Id: I2c9041403dc13435dfcf1db21e1fef25fff0cc2c

Former-commit-id: 1d693d6f0ccd98b45e47c60580662575c2c967cc