Change include compile classpath behavior of annotation processor
Due to performance reasons, all annotation processors must be explicitly
specified.
If includeCompileClasspath is not specified, the plugin will now check the
compile classpath to see if there is any dependencies with annotation
processor.  If one is found, and the dependency is not added to the
annotationProcessor configuration, the build will fail.

Users should manually add the annotation processor to the configuration.
They can also manually set includeCompileClasspath = true to revert to
previous behavior.

Test: Added AnnotationProcessorCompileClasspathTest.
Change-Id: I0e85a5107497e86cba7f77f9b504ee40d95754b0

Former-commit-id: e72dc71b9c6d6aaac9edf14a977b736c022f8307