Fix BuildToolsTest failure.
The test fails because of a typo in TaskStateList.

Once that typo is fixed, there is another failure due to an incorrect
assertion in the test.

The incorrect expected list of up-to-date/outdated tasks was:

:transformClassesWithPreDexForDebug OUTDATED
:transformDexWithDexForDebug OUTDATED
:transformClassesWithPreDexForRelease OUTDATED
:transformDexWithDexForRelease OUTDATED

The correct expected list of up-to-date/outdated tasks should be:

:transformClassesWithDexBuilderForDebug UP-TO-DATE
:transformDexArchiveWithDexMergerForDebug UP-TO-DATE
:transformClassesWithPreDexForRelease OUTDATED
:transformDexWithDexForRelease OUTDATED

(Also note the name changes in the first two tasks of the two lists.)

This commit fixes both issues.

Test: BuildToolsTest (requires manually setting GradleTestProject.
DEFAULT_BUILD_TOOL_VERSION to 25.0.1 for the invalidateBuildTools test
method to run); GradleTaskSubjectTest.hadChangedInputs()

Change-Id: Ib7f37769ce22bd69e47a572e5ac424ee81c5f28f

Former-commit-id: 05d4311d6ed2a13bfd547b6da2baa31c768308aa