[AVD Mgr] Do not create userdata-qemu.img
The AVD Mgr used to copy userdata.img to userdata-qemu.img
when an AVD was created or wiped.

The Emulator no longer likes userdata-qemu.img to be an exact
copy of userdata.img. In particular, userdata-qemu.img needs
some information regarding ADB authorization.

Now the AVD Manager does not create this file at AVD
creation, and deletes this file when wiping. When the
Emulator sees this file missing, it creates it with the
appropriate additional information.

Test: Updated

BUG: b.android.com/255281

Change-Id: I907819d90a90644d0d7cbf10eb1707fc67d862e1

Former-commit-id: 3b44e8ddd90eda7ec8fb013001c8738023d17202