Attempt to fix flaky ddmlib.IntegrationTest
This test seems to fail once in a while on the CI servers. The test expects
a single device but it turns out to have none sometimes. This bug could not
be reproduced locally, so this is an attempt to see if the problem is with
the delay in getting the initial device list.

Test: updated
Change-Id: I668134e255edb51f246e019be9a0bd729181ec24

Former-commit-id: b65620550dc6b4723301733126ced6f14f5e457a