Fix issue with exploding AARs in parallel.
We want to explode AARs in parallel to improve build speed. However, we
should create new threads only for AARs that we know are not yet
extracted.

The previous implementation was an incorrect way of achieving this goal,
resulting in AARs possibly not being exploded after returning from an
API call to the cache (FileCache). This commit fixes this issue.

Test: FileCacheTest, and manually retry building
ub-supportlib-flatfoot-master

Change-Id: Idf7f7f65f5ba50f8031b444d1e8cd6ce1ebdefee

Former-commit-id: 6bc94b4ccb6e1f69c47f8209057d0865d3060372