Stress invoke-virtual wrap instrumentation
This is another instrumentation test case. With this we have all the
mechanics required for the most common instrumentation cases:
- entry hooks (forwarding the incoming arguments)
- exit hooks (and potentially changing the return value)
- wrapping virtual-invokes

For every method body in the .dex image, replace invoke-virtual[/range]
instances with a invoke-static[/range] to a fictitious Tracer.WrapInvoke(<args...>)
WrapInvoke() is a static method which takes the same arguments as the
original method plus an explicit "this" argument, and returns the same
type as the original method.

Test: bazel test :dexter_tests

Change-Id: I53c7ba08bb2e690e18b1f2f18c39debad38f0ed9

Former-commit-id: 1a9262d9debc2c33388d51094028f2be3d1e97d1