Manifest merger: remove tools: namespace elements
The manifest merger already had the capability to
strip out tools: attributes at build time. This CL
adds the ability to also remove *elements* in the
tools namespace. This is useful for example for the
new tools:validation elements used in the manifest
to leave testing URLs for app link verification.

Test: Unit test included
Change-Id: I8d48c75ed6e1d4a433e58e30e3312b3f6ca592e2

Former-commit-id: 50e3ad110f1bcd8f795727040ecc0eb9ab3e7285