Index prototypes and method definitions
Builder::FindMethod() used to do a linear search, which works
for a single class IR or small number of searches but it can easily
become a bottleneck.

We already have built a specialized hash table for indexing IR nodes
(StringsHashTable), so this change we're using it for indexing
prototypes and method definitions as well. In the process the hash table
is made more generic and reusable (and renamed to slicer::HashTable)

Bug: 37428380

Test: bazel test :dexter_tests

Change-Id: Id2e92342e81ae685bbab169f0198649b7f48e68c

Former-commit-id: 0408542a5fb7a8868e13a11e752b59962d3f5d23