Fix handling of updates in IncrementalRelativeFileSets.
IncrementalRelativeFileSets.makeFromBaseFiles() requires information
about whether a deleted path is a file or a directory. Currently, Gradle
does not provide this information. Therefore, we shift the decision on
how to treat a deleted path to the callers of this method (via the
fileDeletionPolicy parameter).

Once Gradle provides this information, we should remove this parameter.

Test: IncrementalRelativeFileSetsTest,
      LazyIncrementalFileMergerInputTest

Change-Id: I15345fe1dc388f3f6b90abb38bd4134455d31f6c

Former-commit-id: 2c52e237385aab877a61425d782dbf3944dceb08