Update lint's findViewById detector to handle O SDK implicit casts
The findViewById signature was updated in Java 8 to implicitly
cast to the type of the LHS. This makes code easier to write, but
it means that the lint check (which was checking the explicit cast)
was no longer enforcing consistency between the type and the tags
used in the corresponding layout.

This CL updates lint to handle this scenario correctly; it now
checks the type of the surrounding expression and uses that inferred
type to perform the compatibility check. It also cleans up the code
a bit and switches from a DOM parser to a pull parser, etc.

It also adds a new check to warn about scenarios where the cast is
incompatible in 1.8 (and adds a quickfix to add the missing cast.)

This CL also adds some lint infrastructure improvements needed for
this fix: first, quickfix descriptors can now request format and
symbol shortening, and the gradle model mocking can request language
levels.

Tests: Unit tests included

Change-Id: If00601dc73cd49310bb7c2eee39de8cf16c8c595

Former-commit-id: b5fb2334be718936028a50ebfc7d8270133f201d