change WaitableExecutor to use ForkJoinPool instead of plain Executor.
This should reduce the chances of seeing deadlocks. A proper fix will
be however to completely get rid of the WaitableExecutor and replace it
with direct ForkJoinPool calls.

Changed the DexTransfrom and below classes to use the ForkJoinPool as
they spawn new tasks while executing work items which could lead to
thread exhaustion with the Executor.

Changed JNI Aapt2 link() to not spawn a new task and just wait for it.
This will work around the bug mentioned below, but again, in a later
CL we should probably use ForkJoinPool.

Bug=37490256
Test: no new test, adapted existing ones.

Change-Id: Ic6f5a786d63782fd2f9ee6025645b8fe1077f999

Former-commit-id: f135807ceaee816a8890103c44db3eb83cfc6694