Enable Level3.
The plugin now requires a Studio that send a level3 model
query. This is to ensure that only Studio 3.0 opens projects
using plugin 3.0 due to the lack of variant output in the
main import sync (variant output info is sent in a new
post-build sync).

The model still currently sends variant output, this needs
to be cleaned up in a separate CLs that will impact of lot
of tests.

This change also introduce a level 4 syncs which is the new
dependency model. This is basically shifting the old level 2
to level 4 so that it's after level 3, as level 2 was after level 1.
This is because it's not ready yet, but hopefully it can be merged
back into level 3 shortly.

Test: new OlderStudioModel that verifies the plugin refuses older studio.
Change-Id: I259ec9897564e37260c91c2bbeb6d0ed734d4f3e

Former-commit-id: 82e51a88662295f33af87846f28098d2310781ff