Add library resource verification in release mode.
With the new resource processing in libraries we skip compiling and
linking with AAPT, instead creating the R.java files ourselves to
improve perfmance. Because of this though, we cannot completely verify
the correctness of the resources (we still check types and values, but
not references).
To fix this, we will now call AAPT compile and link even in libraries,
but only in release mode.

BUG=37139303
Test: included LibWithResourcesTest

Change-Id: Ia29c7694a2dce10f0b18b96c1c80b34322c97ad4

Former-commit-id: 7aec371b9c4d9722b2530f1562cbdd88cfcad466