Fix wrap okhttp response input stream.
okhttp response stream was buffered for a requested length which needs
fix, because it may not buffer all of bytes. If the length was unknown,
things go wrong too. Instead, we create a new object of response that
wraps up the original with tracked stream. When users invoke the wraped
object, the stream bytes will be tracked.

Bug: 37933664

Test: Manually
Change-Id: I9cfe98fec735a77abaacf4d9bd00c8ec1768d31e

Former-commit-id: f03701b295b3540c22bff47f3a9f6c7fd63d0b02