[AIA] Introduce sdkInt in ApkVersionInfo in manifest
Instant app sdk manifest has changed: apks may be different for
different API levels of the device, so a new tag sdkInt was introduced
in ApkVersionInfo node in the manifest. We also parse the
aiaCompatApiMinVersion flag.

This CL adds it to the manifest parser and consumes it. It also does a
minor refactoring moving classes related to parsing the manifest to a
different package.

Test: Unit tests

Bug: Studio part of b/38039710
Change-Id: I2eb548baf7cdde7c93a3323b7ac627f5c3b0a249

Former-commit-id: b6a19db431b147a9fd40620cdd37bda286e93814